+++ CHANGELOG.md
@@ -6,6 +6,7 @@ ChangeLog
 ------------------
 
 - IMPROVEMENT: add "keywords" parameter to node "a2t-deepgram"
+- UPDATE: upgrade NPM dependencies
 
 2.0.4 (2026-01-16)
 ------------------